[libc] Reorganize iswctype to avoid GPU/libc++ error
After #191178 there were build errors when building the libc++
hand-in-hand pieces due to header layering.

Written with the assistance of Gemini
