+++ libc/src/wctype/iswctype.cpp
@@ -7,11 +7,10 @@
 //===----------------------------------------------------------------------===//
 
 #include "src/wctype/iswctype.h"
-#include "src/__support/common.h"
-#include "src/__support/wctype_utils.h"
-
 #include "hdr/types/wctype_t.h"
 #include "hdr/types/wint_t.h"
+#include "src/__support/common.h"
+#include "src/__support/wctype_impl.h"
 
 namespace LIBC_NAMESPACE_DECL {
